+++ CHANGES.rst
@@ -17,7 +17,7 @@ Summary
 +++++++
 
 - Coming soon (November 2017)
-- 16 contributors (7 new)
+- 17 contributors (8 new)
 - 7 months of work (from April 28, 2017 to November 2017)
 - 120 pull requests (not all listed below)
 
@@ -80,6 +80,7 @@ Summary
 - Matthias Wegen (new)
 - Peter Deiml (new)
 - Roberta Zanin (new)
+- Thomas Vuillaume (new)
 - Rubén López-Coto (new)
 
 Pull requests
@@ -89,6 +90,7 @@ This list is incomplete. Small improvements and bug fixes are not listed here.
 
 See the complete `Gammapy 0.7 merged pull requests list on Github <https://github.com/gammapy/gammapy/pulls?utf8=%E2%9C%93&q=is%3Apr+milestone%3A0.7+is%3Amerged+>`__.
 
+- [#1207] Add theta2 distribution plot to EventList class (Thomas Vuillaume)
 - [#1190] Refactor gammapy.maps methods for calculating index and coordinate arrays (Matthew Wood)
 - [#1183] Add function to compute background cube (Roberta Zanin and Christoph Deil)
 - [#1179] Fix two bugs in LightCurveEstimator, and improve speed considerably (Lars Mohrmann)